Rename `[p]deletenames` to `[p]modset deletenames`
